refactoring: misc refactoring
