+++ frontend/package.json
@@ -10,19 +10,15 @@
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
     "apexcharts": "^3.35.3",
-    "axios": "^0.27.2",
     "chakra-ui-markdown-renderer": "^4.0.0",
-    "chart.js": "^3.8.0",
     "dotenv": "^16.0.0",
     "easymde": "^2.16.0",
     "framer-motion": "^6.3.2",
-    "i": "^0.3.7",
     "immer": "^9.0.14",
     "npm": "^8.10.0",
     "react": "^18.0.0",
     "react-apexcharts": "^1.4.0",
     "react-beautiful-dnd": "^10.0.4",
-    "react-chartjs-2": "^4.1.0",
     "react-cookie": "^4.1.1",
     "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
